+++ R/count.R
@@ -34,6 +34,7 @@
 #' The function \code{count_df} takes any variable from \code{data} that has an \code{"rsi"} class (created with \code{\link{as.rsi}}) and counts the amounts of S, I and R. The resulting \emph{tidy data} (see Source) \code{data.frame} will have three rows (S/I/R) and a column for each variable with class \code{"rsi"}.
 #'
 #' The function \code{rsi_df} works exactly like \code{count_df}, but adds the percentage of S, I and R.
+#' @inheritSection portion Combination therapy
 #' @source Wickham H. \strong{Tidy Data.} The Journal of Statistical Software, vol. 59, 2014. \url{http://vita.had.co.nz/papers/tidy-data.html}
 #' @seealso \code{\link{portion}_*} to calculate microbial resistance and susceptibility.
 #' @keywords resistance susceptibility rsi antibiotics isolate isolates
@@ -61,8 +62,8 @@
 #' # Since n_rsi counts available isolates, you can
 #' # calculate back to count e.g. non-susceptible isolates.
 #' # This results in the same:
-#' count_IR(septic_patients$AMX)
-#' portion_IR(septic_patients$AMX) * n_rsi(septic_patients$AMX)
+#' count_SI(septic_patients$AMX)
+#' portion_SI(septic_patients$AMX) * n_rsi(septic_patients$AMX)
 #'
 #' library(dplyr)
 #' septic_patients %>%
@@ -76,17 +77,17 @@
 #'
 #' # Count co-resistance between amoxicillin/clav acid and gentamicin,
 #' # so we can see that combination therapy does a lot more than mono therapy.
-#' # Please mind that `portion_S` calculates percentages right away instead.
-#' count_S(septic_patients$AMC)   # S = 1342 (71.4%)
-#' count_all(septic_patients$AMC) # n = 1879
+#' # Please mind that `portion_SI` calculates percentages right away instead.
+#' count_SI(septic_patients$AMC)  # 1433
+#' count_all(septic_patients$AMC) # 1879
 #'
-#' count_S(septic_patients$GEN)   # S = 1372 (74.0%)
-#' count_all(septic_patients$GEN) # n = 1855
+#' count_SI(septic_patients$GEN)  # 1399
+#' count_all(septic_patients$GEN) # 1855
 #'
 #' with(septic_patients,
-#'      count_S(AMC, GEN))         # S = 1660 (92.3%)
-#' with(septic_patients,           # n = 1798
-#'      n_rsi(AMC, GEN))
+#'      count_SI(AMC, GEN))        # 1764
+#' with(septic_patients,
+#'      n_rsi(AMC, GEN))           # 1936
 #'
 #' # Get portions S/I/R immediately of all rsi columns
 #' septic_patients %>%
@@ -99,71 +100,56 @@
 #'   group_by(hospital_id) %>%
 #'   count_df(translate = FALSE)
 #'
-count_R <- function(..., also_single_tested = FALSE) {
+count_R <- function(..., only_all_tested = FALSE) {
   rsi_calc(...,
-           type = "R",
-           include_I = FALSE,
-           minimum = 0,
-           as_percent = FALSE,
-           also_single_tested = also_single_tested,
+           ab_result = "R",
+           only_all_tested = only_all_tested,
            only_count = TRUE)
 }
 
 #' @rdname count
 #' @export
-count_IR <- function(..., also_single_tested = FALSE) {
+count_IR <- function(..., only_all_tested = FALSE) {
   rsi_calc(...,
-           type = "R",
-           include_I = TRUE,
-           minimum = 0,
-           as_percent = FALSE,
-           also_single_tested = also_single_tested,
+           ab_result = c("I", "R"),
+           only_all_tested = only_all_tested,
            only_count = TRUE)
 }
 
 #' @rdname count
 #' @export
-count_I <- function(..., also_single_tested = FALSE) {
+count_I <- function(..., only_all_tested = FALSE) {
   rsi_calc(...,
-           type = "I",
-           include_I = FALSE,
-           minimum = 0,
-           as_percent = FALSE,
-           also_single_tested = also_single_tested,
+           ab_result = "I",
+           only_all_tested = only_all_tested,
            only_count = TRUE)
 }
 
 #' @rdname count
 #' @export
-count_SI <- function(..., also_single_tested = FALSE) {
+count_SI <- function(..., only_all_tested = FALSE) {
   rsi_calc(...,
-           type = "S",
-           include_I = TRUE,
-           minimum = 0,
-           as_percent = FALSE,
-           also_single_tested = also_single_tested,
+           ab_result = c("S", "I"),
+           only_all_tested = only_all_tested,
            only_count = TRUE)
 }
 
 #' @rdname count
 #' @export
-count_S <- function(..., also_single_tested = FALSE) {
+count_S <- function(..., only_all_tested = FALSE) {
   rsi_calc(...,
-           type = "S",
-           include_I = FALSE,
-           minimum = 0,
-           as_percent = FALSE,
-           also_single_tested = also_single_tested,
+           ab_result = "S",
+           only_all_tested = only_all_tested,
            only_count = TRUE)
 }
 
 #' @rdname count
 #' @export
-count_all <- function(..., also_single_tested = FALSE) {
-  res_SI <- count_SI(..., also_single_tested = also_single_tested)
-  # only print warnings once, if needed
-  res_R <- suppressWarnings(count_R(..., also_single_tested = also_single_tested))
-  res_SI + res_R
+count_all <- function(..., only_all_tested = FALSE) {
+  rsi_calc(...,
+           ab_result = c("S", "I", "R"),
+           only_all_tested = only_all_tested,
+           only_count = TRUE)
 }
 
 #' @rdname count